requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
Deprecated==1.2.12
dnspython==1.16.0
email-validator==1.1.3
fastapi==0.68.0
fastapi-users==6.1.2
h11==0.12.0
makefun==1.11.3
motor==2.5.1
passlib==1.7.4
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.12.0
python-dotenv==0.19.0
python-multipart==0.0.5
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.15.0
wrapt==1.12.1
